refactored var types

+++ Cavestory/Sprite.h
@@ -12,7 +12,7 @@ class Sprite
 public:
 	Sprite();
 	Sprite(Graphics &graphics, const std::string &filePath, int sourceX, int sourceY, int width, int height,
-		float posX, float posY);
+		int posX, int posY);
 	virtual ~Sprite();
 	virtual void Update();
 	void Draw(Graphics &graphics, int x, int y);
@@ -21,6 +21,6 @@ private:
 	SDL_Rect _sourceRect;
 	SDL_Texture *_spriteSheet;
 
-	float _x, _y;
+	int _x, _y;
 };
